wifi: update driver to 100.10.545.2 to support STA/AP concurrent [1/2]
PD#SWPL-12561

Problem:
STA/AP concurrent support

Solution:
update driver to fix it

Verify:
ampere

+++ bcmdhd.100.10.315.x/frag.c
@@ -2,7 +2,7 @@
  * IE/TLV fragmentation/defragmentation support for
  * Broadcom 802.11bang Networking Device Driver
  *
- * Copyright (C) 1999-2018, Broadcom.
+ * Copyright (C) 1999-2019, Broadcom.
  *
  *      Unless you and Broadcom execute a separate written software license
  * agreement governing use of this software, this software is licensed to you
@@ -73,7 +73,7 @@ bcm_tlv_dot11_defrag(const void *buf, uint buf_len, uint8 id, bool id_ext,
 	}
 
 	/* adjust buf_len to length after ie including it */
-	buf_len -= ((const uint8 *)ie - (const uint8 *)buf);
+	buf_len -= (uint)(((const uint8 *)ie - (const uint8 *)buf));
 
 	/* update length from fragments, okay if no next ie */
 	while ((ie = bcm_next_tlv(ie, &buf_len)) &&
